% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logo}
\alias{logo}
\title{Tidylab ASCII logo}
\format{
An object of class \code{character} of length 32.
}
\usage{
logo
}
\description{
Tidylab ASCII logo
}
\examples{
print(logo)
}
\keyword{internal}
