% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R, R/verse.R
\name{tidylab_conflicts}
\alias{tidylab_conflicts}
\title{Conflicts between the tidylab and other packages}
\usage{
tidylab_conflicts()

tidylab_conflicts()
}
\value{
(\code{tidylab_conflicts}) A list with conflicted functions.
}
\description{
This function lists all the conflicts between packages in the tidylab
and other packages that you have loaded.

This function lists all the conflicts between packages in the tidylab
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.

If dplyr is one of the select packages, then the following four conflicts
are deliberately ignored: \code{intersect}, \code{union}, \code{setequal},
and \code{setdiff} from dplyr. These functions make the base equivalents
generic, so shouldn't negatively affect any existing code.
}
\examples{
tidylab_conflicts()
tidylab_conflicts()
}
