% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidylab_deps}
\alias{tidylab_deps}
\title{List all tidylab dependencies}
\usage{
tidylab_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidylab packages.}
}
\value{
(\code{tibble}) A table with package dependencies.
}
\description{
List all tidylab dependencies
}
