% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{tidylab_logo}
\alias{tidylab_logo}
\title{The tidylab logo, using ASCII or Unicode characters}
\usage{
tidylab_logo(unicode = l10n_info()$`UTF-8`)
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\value{
(\code{tidylab_logo}) ASCII art of the Tidylab logo.
}
\description{
Use \code{\link[crayon:strip_style]{crayon::strip_style()}} to get rid of the colors.
}
\examples{
tidylab_logo()
}
