% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{tidylab_packages}
\alias{tidylab_packages}
\title{List all packages in the tidylab}
\usage{
tidylab_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include tidylab in the list?}
}
\value{
(\code{character}) The names of the imported packages.
}
\description{
List all packages in the tidylab
}
\examples{
tidylab_packages()
}
